agregue DTO

+++ src/routes/products.router.js
@@ -1,7 +1,9 @@
 import { Router } from 'express';
 import { Server } from "socket.io";
 import productController from '../controllers/product.controller.js'
-import { logger } from '../middlewares/logger.middleware.js'; // Añadimos el import del logger
+import { logger } from '../middlewares/logger.middleware.js'; 
+import { ensurePremiumOrAdmin } from '../middlewares/auth.middleware.js';
+import { middlewarePassportJwt } from '../middlewares/jwt.middleware.js';
 
 const productsRouter = Router();
 
@@ -9,16 +11,18 @@ const server = new Server();
 const io = server.io;
 
 productsRouter.get('/', async (req, res) => {
-	try {
-		const products = await productController.obtenerProductosPaginados();
-		res.send(products);
-	} catch (err) {
-		logger.error("Error al obtener productos paginados: " + err.message);
-		res.status(500).send({ err });
-	}
+    try {
+        const { limit, page, sort, query } = req.query; 
+        const products = await productController.obtenerProductosPaginados(limit, page, sort, query);
+        res.send(products);
+    } catch (err) {
+        logger.error("Error al obtener productos paginados: " + err.message);
+        res.status(500).send({ err });
+    }
 });
 
-productsRouter.post('/',  async (req, res) => {
+
+productsRouter.post('/',middlewarePassportJwt,ensurePremiumOrAdmin,  async (req, res) => {
     try {
         const product = await productController.agregarProducto(req.body);
         res.redirect('/products');
@@ -27,7 +31,7 @@ productsRouter.post('/',  async (req, res) => {
         res.status(500).send({ error: err.message });
     } 
 });
-productsRouter.put('/:pid',  async (req, res) => {
+productsRouter.put('/:pid',middlewarePassportJwt,ensurePremiumOrAdmin,  async (req, res) => {
 	const pid = req.params.pid;
 	try {
 		const product = await productController.actualizarProducto(pid, req.body);
@@ -39,19 +43,27 @@ productsRouter.put('/:pid',  async (req, res) => {
 	}
 });
 
-productsRouter.delete('/:pid', async (req, res) => {
-	const pid = req.params.pid;
-	try {
-		await productController.eliminarProducto(pid);
-		req.io.emit("updatedProducts", await req.productController.obtenerProductos());
-		res.sendStatus(204);
-	} catch (err) {
-		logger.error("Error al eliminar producto: " + err.message);
-		res.status(500).send({ err });
-	}
+productsRouter.delete('/:pid',middlewarePassportJwt, ensurePremiumOrAdmin, async (req, res) => {
+    const pid = req.params.pid;
+    try {
+        const product = await productController.obtenerProductoById(pid);
+
+        if (req.user.rol !== 'ADMIN' && String(req.user._id) !== String(product.owner)) {
+            logger.warn(`Usuario con ID ${req.user._id} intentó eliminar un producto que no le pertenece`);
+            return res.status(403).send({ message: "No tienes permisos para eliminar este producto" });
+        }
+
+        await productController.eliminarProducto(pid);
+        req.io.emit("updatedProducts", await req.productController.obtenerProductos());
+        res.sendStatus(204);
+    } catch (err) {
+        logger.error("Error al eliminar producto: " + err.message);
+        res.status(500).send({ err });
+    }
 });
 
-productsRouter.get('/:pid', async (req, res) => {
+
+productsRouter.get('/:pid',middlewarePassportJwt,ensurePremiumOrAdmin, async (req, res) => {
 	try {
 		let productoSolicitado = await productController.obtenerProductoById(req.params.pid);
 		res.send({ producto: productoSolicitado });
